pyproject.toml
[project]
name = "taboo"
version = "0.1.0"

dependencies = [
    "python-dotenv==1.1.0",
    "openai==1.75.0",
    "numpy==2.2.5",
    "tqdm==4.67.1"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core", "agent"]